+++ shared/src/admin/admin.routes.ts
@@ -11,7 +11,7 @@ import { setFeatureFlagSchema } from "../featureFlag/featureFlags.schema";
 import {
   establishmentBatchReportSchema,
   formEstablishmentBatchSchema,
-} from "../formEstablishment/FormEstablishment.schema";
+} from "../formEstablishment/FormEstablishments.schema";
 import { withAuthorizationHeaders } from "../headers";
 import {
   httpErrorSchema,
