+++ shared/src/index.ts
@@ -51,9 +51,9 @@ export * from "./errors/errors";
 export * from "./featureFlag/featureFlags.dto";
 export * from "./featureFlag/featureFlags.schema";
 export * from "./federatedIdentities/federatedIdentity.dto";
-export * from "./formEstablishment/FormEstablishment.dto";
-export * from "./formEstablishment/FormEstablishment.schema";
-export * from "./formEstablishment/FormEstablishmentDtoBuilder";
+export * from "./formEstablishment/FormEstablishments.dto";
+export * from "./formEstablishment/FormEstablishments.schema";
+export * from "./formEstablishment/FormEstablishmentsDtoBuilder";
 export * from "./geoPosition/geoPosition.dto";
 export * from "./geoPosition/geoPosition.schema";
 export * from "./group/group.dto";
